Feature dev/lyq

See merge request ckb-explorer/ckb-explorer-web!39

+++ src/main/java/com/ckb/explorer/domain/resp/AddressBalanceRanking.java
@@ -2,6 +2,7 @@ package com.ckb.explorer.domain.resp;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
+import java.math.BigInteger;
 import lombok.Data;
 
 @Data
@@ -10,7 +11,7 @@ public class AddressBalanceRanking {
   private String address;
 
   @JsonSerialize(using = ToStringSerializer.class)
-  private Long balance;
+  private BigInteger balance;
 
   private int ranking;
 }
